requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorlog==6.6.0
cryptography==36.0.1
falcon==3.0.1
idna==3.3
passlib==1.7.4
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
urllib3==1.26.8
